requirements.txt
amanobot
bs4
colorama
requests
youtube-dl
duckpy
python-dotenv
